[minor] MoveGuard now lives in its own file.
